chore: move various types to shared-lib
