+++ meteor/server/api/ingest/mosDevice/actions.ts
@@ -12,7 +12,7 @@ import * as _ from 'underscore'
 import { TriggerReloadDataResponse } from '../../../../lib/api/userActions'
 import { runIngestOperation } from '../lib'
 import { IngestJobs } from '@sofie-automation/corelib/dist/worker/ingest'
-import { DEFAULT_MOS_TIMEOUT_TIME } from '@sofie-automation/shared-lib'
+import { DEFAULT_MOS_TIMEOUT_TIME } from '@sofie-automation/shared-lib/dist/core/constants'
 
 export namespace MOSDeviceActions {
 	export async function reloadRundown(
