+++ packages/blueprints-integration/package.json
@@ -39,6 +39,7 @@
 	],
 	"dependencies": {
 		"timeline-state-resolver-types": "7.1.0-release42.1",
+		"@sofie-automation/shared-lib": "1.44.0-in-development",
 		"tslib": "^2.4.0",
 		"type-fest": "^2.12.2"
 	},
