+++ packages/corelib/src/dataModel/PeripheralDeviceSettings/ingestDevice.ts
@@ -1,28 +1,6 @@
-export interface IngestDeviceSettings {
-	/** OAuth: Set to true when secret value exists */
-	secretCredentials: boolean
-	secretAccessToken: boolean
-}
-export interface IngestDeviceSecretSettings {
-	/** OAuth: */
-	credentials?: Credentials
-	accessToken?: AccessToken
-}
-export interface Credentials {
-	installed: {
-		client_id: string
-		project_id: string
-		auth_uri: string
-		token_uri: string
-		auth_provider_x509_cert_url: string
-		client_secret: string
-		redirect_uris: string[]
-	}
-}
-export interface AccessToken {
-	access_token: string
-	refresh_token: string
-	scope: string
-	token_type: string
-	expiry_date: number
-}
+export * from '@sofie-automation/shared-lib/dist/peripheralDevice/ingest'
+
+export {
+	IngestDeviceSettings,
+	IngestDeviceSecretSettings,
+} from '@sofie-automation/shared-lib/dist/core/model/peripheralDevice'
