+++ packages/corelib/src/dataModel/PeripheralDeviceSettings/playoutDevice.ts
@@ -1,14 +1,1 @@
-import { TSR } from '@sofie-automation/blueprints-integration'
-
-/**
- * The basic PlayoutDevice settings structure.
- * Note: playout-gateway will likely have more than this here, but this is that core needs to know about
- */
-export interface PlayoutDeviceSettings {
-	devices: {
-		[deviceId: string]: TSR.DeviceOptionsAny
-	}
-	locations: {
-		[deviceId: string]: any // todo
-	}
-}
+export * from '@sofie-automation/shared-lib/dist/peripheralDevice/ingest'
