+++ packages/job-worker/src/events/handle.ts
@@ -25,7 +25,7 @@ import {
 } from '@sofie-automation/corelib/dist/dataModel/PeripheralDevice'
 import { MOS } from '@sofie-automation/corelib'
 import { executePeripheralDeviceFunction } from '../peripheralDevice'
-import { DEFAULT_MOS_TIMEOUT_TIME } from '@sofie-automation/shared-lib'
+import { DEFAULT_MOS_TIMEOUT_TIME } from '@sofie-automation/shared-lib/dist/core/constants'
 
 async function getBlueprintAndDependencies(context: JobContext, rundown: ReadonlyDeep<DBRundown>) {
 	const pShowStyle = context.getShowStyleCompound(rundown.showStyleVariantId, rundown.showStyleBaseId)
