+++ packages/mos-gateway/src/configManifest.ts
@@ -3,7 +3,10 @@
  * UI.
  */
 import { DeviceConfigManifest, ConfigManifestEntryType } from '@sofie-automation/server-core-integration'
-import { DEFAULT_MOS_TIMEOUT_TIME, DEFAULT_MOS_HEARTBEAT_INTERVAL } from '@sofie-automation/shared-lib'
+import {
+	DEFAULT_MOS_TIMEOUT_TIME,
+	DEFAULT_MOS_HEARTBEAT_INTERVAL,
+} from '@sofie-automation/shared-lib/dist/core/constants'
 
 export const MOS_DEVICE_CONFIG_MANIFEST: DeviceConfigManifest = {
 	deviceConfig: [
