+++ packages/server-core-integration/package.json
@@ -70,6 +70,7 @@
 		"production"
 	],
 	"dependencies": {
+		"@sofie-automation/shared-lib": "1.44.0-in-development",
 		"data-store": "^4.0.3",
 		"ejson": "^2.2.2",
 		"eventemitter3": "^4.0.7",
