+++ packages/shared-lib/src/index.ts
@@ -1,1 +1,2 @@
-export * from './core/constants'
+// No exports here, import directly instead:
+// import * from '@sofie-automation/shared-lib/dist/lib/lib'
