+++ packages/shared-lib/src/lib/lib.ts
@@ -1,0 +1,2 @@
+export type Time = number
+export type TimeDuration = number
