+++ packages/shared-lib/src/peripheralDevice/ingest.ts
@@ -1,0 +1,4 @@
+export type IngestPlaylist = any
+export type IngestRundown = any
+export type IngestSegment = any
+export type IngestPart = any
