+++ packages/shared-lib/src/peripheralDevice/mos.ts
@@ -1,0 +1,15 @@
+// Faking the MOS interface for now, so we don't have to expose mos-connection to the client
+
+export type IMOSItem = any
+export type IMOSItemAction = any
+export type IMOSItemStatus = any
+export type IMOSROAction = any
+export type IMOSROFullStory = any
+export type IMOSROReadyToAir = any
+export type IMOSROStory = any
+export type IMOSRunningOrder = any
+export type IMOSRunningOrderBase = any
+export type IMOSRunningOrderStatus = any
+export type IMOSStoryAction = any
+export type IMOSStoryStatus = any
+export type MosString128 = any
